+++ mrv2/lib/mrvCore/mrvHotkey.h
@@ -267,6 +267,7 @@ namespace mrv
     extern Hotkey kEditCutFrame;
     extern Hotkey kEditCopyFrame;
     extern Hotkey kEditPasteFrame;
+    extern Hotkey kEditInsertFrame;
 
     // @todo:
     extern Hotkey kRotatePlus90;