+++ mrv2/lib/mrvFl/mrvCallbacks.cpp
@@ -9,18 +9,6 @@ namespace fs = std::filesystem;
 #include <fstream>
 #include <iomanip>
 
-#include <opentime/rationalTime.h>
-#include <opentime/timeRange.h>
-
-#include <opentimelineio/clip.h>
-#include <opentimelineio/composition.h>
-#include <opentimelineio/editAlgorithm.h>
-#include <opentimelineio/gap.h>
-#include <opentimelineio/item.h>
-#include <opentimelineio/mediaReference.h>
-#include <opentimelineio/timeline.h>
-#include <opentimelineio/transition.h>
-
 #include <tlCore/StringFormat.h>
 
 #include <FL/filename.H> // for fl_open_uri()
@@ -70,6 +58,23 @@ namespace
 {
     const char* kModule = "cback";
 }
+
+namespace
+{
+    using namespace tl;
+
+    void debug_composition_children(otio::Composition* composition)
+    {
+        for (auto child : composition->children())
+        {
+            auto clip = otio::dynamic_retainer_cast<otio::Clip>(child);
+            if (!clip)
+                continue;
+            std::cout << clip->name() << " " << clip->trimmed_range()
+                      << std::endl;
+        }
+    }
+} // namespace
 
 namespace mrv
 {
@@ -676,29 +681,6 @@ namespace mrv
          has_bottom_bar = true, has_pixel_bar = true, has_status_bar = true,
          has_dock_grp = false, has_preferences_window = false,
          has_hotkeys_window = false, has_about_window = false;
-    EditMode editMode = EditMode::kTimeline;
-    int editModeH = 30;
-    int kMinEditModeH = 30;
-
-    void save_edit_mode_state(ViewerUI* ui)
-    {
-        int H = ui->uiTimelineGroup->h();
-
-        if (H == 0)
-        {
-            editMode = EditMode::kNone;
-        }
-        else if (H > kMinEditModeH)
-        {
-            editMode = EditMode::kSaved;
-            editModeH = H;
-        }
-        else
-        {
-            editMode = EditMode::kTimeline;
-            editModeH = kMinEditModeH;
-        }
-    }
 
     void save_ui_state(ViewerUI* ui, Fl_Group* bar)
     {
@@ -1835,286 +1817,6 @@ namespace mrv
         auto path = item->path.get();
 
         file_manager_show_uri(path);
-    }
-
-    void set_edit_mode_cb(EditMode mode, ViewerUI* ui)
-    {
-        Fl_Button* b = ui->uiEdit;
-
-        bool active = (mode == EditMode::kFull || mode == EditMode::kSaved);
-        if (mode == EditMode::kSaved && editModeH == kMinEditModeH)
-            active = false;
-
-        b->value(active);
-        if (b->value())
-        {
-            b->labelcolor(fl_rgb_color(255, 255, 255));
-        }
-        else
-        {
-            b->labelcolor(FL_FOREGROUND_COLOR);
-        }
-        b->redraw();
-
-        Fl_Tile* tile = ui->uiTileGroup;
-        Fl_Group* timeline = ui->uiTimelineGroup;
-        Fl_Flex* view = ui->uiViewGroup;
-        int tileY = tile->y();
-        int oldY = timeline->y();
-        int timelineH = timeline->h();
-        int tileH = tile->h();
-        int H = kMinEditModeH; // timeline height
-        int viewH = H;
-        auto player = ui->uiView->getTimelinePlayer();
-        if (mode == EditMode::kFull && player)
-        {
-            // Shift the view up to see the video thumbnails and audio waveforms
-            const double pixelRatio = ui->uiTimeline->pixels_per_unit();
-            const int maxTileHeight = tileH - 20;
-            const timelineui::ItemOptions options =
-                ui->uiTimeline->getItemOptions();
-            auto otioTimeline = player->timeline()->getTimeline();
-            for (const auto& child : otioTimeline->tracks()->children())
-            {
-                if (const auto* track = dynamic_cast<otio::Track*>(child.value))
-                {
-                    if (otio::Track::Kind::video == track->kind())
-                    {
-                        H += 24; // title bar
-                        if (options.thumbnails)
-                            H += options.thumbnailHeight / pixelRatio;
-                        H += 24; // bottom bar
-                    }
-                    else if (otio::Track::Kind::audio == track->kind())
-                    {
-                        H += 24; // title bar
-                        if (options.thumbnails)
-                            H += options.waveformHeight / pixelRatio;
-                        H += 24; // bottom bar
-                    }
-                    // Handle transitions
-                    if (options.showTransitions)
-                    {
-                        bool found = false;
-                        for (const auto& child : track->children())
-                        {
-                            if (const auto& transition =
-                                    dynamic_cast<otio::Transition*>(
-                                        child.value))
-                            {
-                                found = true;
-                                break;
-                            }
-                        }
-                        if (found)
-                            H += 20;
-                    }
-                }
-            }
-
-            if (H >= maxTileHeight)
-                H = maxTileHeight;
-
-            editMode = EditMode::kSaved;
-            editModeH = viewH = H;
-            timeline->show();
-            if (ui->uiMain->visible())
-                ui->uiTimeline->show();
-        }
-        else if (mode == EditMode::kSaved)
-        {
-            H = viewH = editModeH;
-            timeline->show();
-            if (ui->uiMain->visible())
-                ui->uiTimeline->show();
-        }
-        else if (mode == EditMode::kNone)
-        {
-            viewH = 0;
-            ui->uiTimeline->hide();
-            timeline->hide();
-        }
-        else
-        {
-            H = kMinEditModeH; // timeline height
-            viewH = editModeH = H;
-
-            // EditMode::kTimeline
-            timeline->show();
-            if (ui->uiMain->visible())
-                ui->uiTimeline->show();
-        }
-
-        int newY = tileY + tileH - H;
-
-        view->resize(view->x(), view->y(), view->w(), tileH - viewH);
-        if (timeline->visible())
-            timeline->resize(timeline->x(), newY, timeline->w(), H);
-
-        if (mode != EditMode::kNone)
-        {
-            assert(view->h() + timeline->h() == tile->h());
-            assert(timeline->y() == view->y() + view->h());
-            tcp->pushMessage("setEditMode", (int)mode);
-        }
-
-        view->layout();
-        tile->init_sizes();
-
-        if (timeline->visible())
-            timeline->redraw(); // needed
-
-        // std::cerr << "editModeH=" << editModeH << std::endl;
-        // std::cerr << "tileY=" << tileY << std::endl;
-        // std::cerr << "tileH=" << tileH << " tileMY=" << tileY + tileH
-        //           << std::endl
-        //           << std::endl;
-        // std::cerr << "viewgroupX=" << view->x() << std::endl;
-        // std::cerr << "viewgroupW=" << view->w()
-        //           << " viewgroupMX=" << view->x() + view->w() << std::endl;
-        // std::cerr << "uiToolGroupX=" << ui->uiToolsGroup->x() << std::endl;
-        // std::cerr << "uiToolGroupW=" << ui->uiToolsGroup->w()
-        //           << " uiToolsGroupMX="
-        //           << (ui->uiToolsGroup->x() + ui->uiToolsGroup->w())
-        //           << std::endl;
-        // std::cerr << "uiViewX=" << ui->uiView->x() << std::endl;
-        // std::cerr << "uiViewW=" << ui->uiView->w()
-        //           << " uiViewMX=" << (ui->uiView->x() + ui->uiView->w())
-        //           << std::endl;
-        // std::cerr << "uiDockGroupX=" << ui->uiDockGroup->x() << std::endl;
-        // std::cerr << "uiDockGroupW=" << ui->uiDockGroup->w()
-        //           << " uiDockGroupMX="
-        //           << (ui->uiDockGroup->x() + ui->uiDockGroup->w())
-        //           << std::endl
-        //           << std::endl;
-        // std::cerr << "viewgroupY=" << view->y() << std::endl;
-        // std::cerr << "viewgroupH=" << view->h()
-        //           << " viewgroupMY=" << view->y() + view->h() << std::endl;
-        // std::cerr << "uiViewY=" << ui->uiView->y() << std::endl;
-        // std::cerr << "uiViewH=" << ui->uiView->h()
-        //           << " uiViewMY=" << (ui->uiView->y() + ui->uiView->h())
-        //           << std::endl;
-        // std::cerr << "timelineGroupY=" << timeline->y() << std::endl;
-        // std::cerr << "timelineGroupH=" << timeline->h()
-        //           << " timelineGroupMY=" << (timeline->y() + timeline->h())
-        //           << std::endl;
-        // std::cerr << "uiTimelineX=" << ui->uiTimeline->x() << std::endl;
-        // std::cerr << "uiTimelineY=" << ui->uiTimeline->y() << std::endl;
-        // std::cerr << "uiTimelineH=" << ui->uiTimeline->h()
-        //           << " uiTimelineMY="
-        //           << (ui->uiTimeline->y() + ui->uiTimeline->h())
-        //           << std::endl;
-        // std::cerr << std::endl;
-    }
-
-    struct FrameInfo
-    {
-        otio::Composition* composition;
-        std::string type;
-        otio::Item* item;
-        otio::RationalTime time;
-    };
-
-    static std::vector<FrameInfo> bufferedFrames;
-
-    void edit_copy_frame_cb(Fl_Menu_* m, ViewerUI* ui)
-    {
-        auto player = ui->uiView->getTimelinePlayer();
-        if (!player)
-            return;
-
-        auto timeline = player->getTimeline();
-        const auto time = player->currentTime();
-
-        otio::ErrorStatus errorStatus;
-        otime::TimeRange range(time, otime::RationalTime(1.0, time.rate()));
-        auto items = timeline->find_children<otio::Item>(&errorStatus, range);
-        if (items.empty())
-        {
-            return;
-        }
-
-        bufferedFrames.clear();
-        for (const auto& item : items)
-        {
-            otio::Track* track = dynamic_cast<otio::Track*>(item->parent());
-            if (!track)
-                continue;
-
-            FrameInfo frame;
-            frame.composition = track;
-            frame.type = track->kind();
-            auto clonedItem = item->clone();
-            auto range = item->trimmed_range();
-            range = otime::TimeRange(
-                range.start_time(), otime::RationalTime(1.0, time.rate()));
-            otio::Clip* clip = dynamic_cast<otio::Clip*>(clonedItem);
-            if (clip)
-            {
-                clip->set_source_range(range);
-                frame.item = clip;
-            }
-            else
-            {
-                otio::Gap* gap = dynamic_cast<otio::Gap*>(clonedItem);
-                if (!gap)
-                    continue;
-                gap->set_source_range(range);
-                frame.item = gap;
-            }
-            frame.time = time;
-            bufferedFrames.push_back(frame);
-        }
-    }
-
-    void edit_cut_frame_cb(Fl_Menu_* m, ViewerUI* ui)
-    {
-        auto players = ui->uiView->getTimelinePlayers();
-        if (players.empty())
-            return;
-
-        auto player = players[0];
-        auto timeline = player->getTimeline();
-
-        edit_copy_frame_cb(m, ui);
-
-        const auto time = player->currentTime();
-
-        player->setTimeline(nullptr);
-        for (const auto& frame : bufferedFrames)
-        {
-            otio::Composition* composition = frame.composition;
-            otio::algo::slice(composition, time);
-            int index = composition->index_of_child(frame.item);
-            otime::RationalTime out_time =
-                time + otime::RationalTime(1.0, time.rate());
-            otio::algo::slice(composition, out_time);
-            composition->remove_child(index - 1);
-        }
-        player->setTimeline(timeline);
-
-        // Set the end frame in the
-        auto startFrame = timeline->global_start_time().value();
-        auto endFrame = startFrame + timeline->duration();
-        TimelineClass* c = ui->uiTimeWindow;
-        c->uiEndFrame->setTime(endFrame);
-    }
-
-    void edit_paste_frame_cb(Fl_Menu_* m, ViewerUI* ui)
-    {
-        auto player = ui->uiView->getTimelinePlayer();
-        if (!player)
-            return;
-
-        auto timeline = player->getTimeline();
-        const auto time = player->currentTime();
-
-        player->setTimeline(nullptr);
-        for (const auto& frame : bufferedFrames)
-        {
-            otio::algo::insert(frame.item, frame.composition, time);
-        }
-        player->setTimeline(timeline);
     }
 
 } // namespace mrv