+++ mrv2/lib/mrvFl/mrvPreferences.cpp
@@ -1487,10 +1487,39 @@ namespace mrv
             ui->uiPixelBar->hide();
         }
 
+        //
+        // Edit mode options
+        //
+        auto options = ui->uiTimeline->getItemOptions();
+        options.showTransitions = uiPrefs->uiPrefsShowTransitions->value();
+
+        int thumbnails = uiPrefs->uiPrefsEditThumbnails->value();
+        options.thumbnails = true;
+        switch (thumbnails)
+        {
+        case 0:
+            options.thumbnails = false;
+            break;
+        case 1: // Small
+            options.thumbnailHeight = 100;
+            break;
+        case 2: // Medium
+            options.thumbnailHeight = 200;
+            break;
+        case 3: // Large
+            options.thumbnailHeight = 300;
+            break;
+        }
+        options.waveformHeight = options.thumbnailHeight / 2;
+        ui->uiTimeline->setItemOptions(options);
+
         if (uiPrefs->uiPrefsTimeline->value())
         {
             ui->uiBottomBar->show();
-            set_edit_mode_cb(EditMode::kSaved, ui);
+            if (ui->uiEdit->value())
+                set_edit_mode_cb(EditMode::kFull, ui);
+            else
+                set_edit_mode_cb(EditMode::kSaved, ui);
         }
         else
         {
@@ -1688,34 +1717,6 @@ namespace mrv
                 window->always_on_top(value);
             }
         }
-
-        //
-        // Edit mode options
-        //
-        auto options = ui->uiTimeline->getItemOptions();
-        options.showTransitions = uiPrefs->uiPrefsShowTransitions->value();
-
-        int thumbnails = uiPrefs->uiPrefsEditThumbnails->value();
-        options.thumbnails = true;
-        switch (thumbnails)
-        {
-        case 0:
-            options.thumbnails = false;
-            break;
-        case 1: // Small
-            options.thumbnailHeight = 100;
-            break;
-        case 2: // Medium
-            options.thumbnailHeight = 200;
-            break;
-        case 3: // Large
-            options.thumbnailHeight = 300;
-            break;
-        }
-        options.waveformHeight = options.thumbnailHeight / 2;
-        ui->uiTimeline->setItemOptions(options);
-        if (ui->uiEdit->value())
-            set_edit_mode_cb(EditMode::kFull, ui);
 
         ui->uiMain->fill_menu(ui->uiMenuBar);