+++ mrv2/docs/HISTORY.md
@@ -18,6 +18,14 @@ v0.7.2
   were no transitions in the timeline.
 - Added Timeline Preferences to show thumbnails, transitions and markers.
 - Fixed pixel aspect ratio of saved OpenEXR images when they were not 1.
+- Made Window resizing take into account Editing Viewport at start up.
+- Fixed Log Panel when an error was shown to resize to the size of the window
+  and not smaller.
+- Log Panel will no longer open when the file requester is open. 
+- Added an Edit/Frame/Cut, Edit/Frame/Copy, Edit/Frame/Paste and
+  Edit/Frame/Insert to cut, copy, paste and insert one frame (video and audio)
+  of any media.  Currently, it does not support transitions, that are removed.
+
 
 v0.7.1
 ------